Create styles directory for css files and moved all existing ones there.
