+++ step-capstone/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './App.css';
+import './styles/App.css';
 import TravelObject from './components/TravelObject'
 
 class App extends React.Component {
